Agregando salida de excepciones, verificando si el usuario existe  antes de realizar cualquier operacion

+++ src/main/java/com/upc/usuario_gestor/controllers/UsuarioController.java
@@ -6,6 +6,7 @@ import com.upc.usuario_gestor.DTO.NotaDTO;
 import com.upc.usuario_gestor.DTO.UsuarioDto;
 import com.upc.usuario_gestor.entities.Usuario;
 import com.upc.usuario_gestor.exceptions.UsuarioGestorExceptions;
+import com.upc.usuario_gestor.repositories.UsuarioRepository;
 import com.upc.usuario_gestor.responses.UsuarioGestorResponse;
 import com.upc.usuario_gestor.services.UsuarioService;
 import io.swagger.annotations.ApiOperation;
@@ -22,13 +23,23 @@ public class UsuarioController
     @Autowired
     private UsuarioService usuarioService;
 
+    @Autowired
+    private UsuarioRepository usuarioRepository;
+
     @ResponseStatus(HttpStatus.OK)
     @ApiOperation("Devuelve la información de un Usuario, atraves de su identificador.")
     @GetMapping("/usuarios/{usuarioId}")
     public UsuarioGestorResponse<UsuarioDto> getUsuarioById(@PathVariable Long usuarioId)
             throws UsuarioGestorExceptions {
-       return new UsuarioGestorResponse<>("Succes",String.valueOf(HttpStatus.OK),"OK",
-                usuarioService.getUsuarioById(usuarioId));
+        if(usuarioRepository.existsById(usuarioId))
+        {
+            return new UsuarioGestorResponse<>("Succes",String.valueOf(HttpStatus.OK),"OK",
+                    usuarioService.getUsuarioById(usuarioId));
+        }else
+            {
+                return new UsuarioGestorResponse<>("Fallo al Obtener Usuario",String.valueOf(HttpStatus.BAD_REQUEST),"Usuario no existente");
+            }
+
     }
 
     @ResponseStatus(HttpStatus.OK)
@@ -45,8 +56,13 @@ public class UsuarioController
     @PostMapping("/usuarios")
     public UsuarioGestorResponse<UsuarioDto> createUsuario(@RequestBody CreateUsuarioDto createUsuarioDto)
             throws UsuarioGestorExceptions{
-        return new UsuarioGestorResponse<>("Succes",String.valueOf(HttpStatus.OK),"OK",
-                usuarioService.createUsuario(createUsuarioDto));
+        if(usuarioRepository.existsUsuarioByCorreo(createUsuarioDto.getCorreo()))
+        {
+            return new UsuarioGestorResponse<>("Fallo al crear Usuario",String.valueOf(HttpStatus.BAD_REQUEST),"El correo ya esta registrado");
+        }else {
+            return new UsuarioGestorResponse<>("Succes", String.valueOf(HttpStatus.OK), "OK",
+                    usuarioService.createUsuario(createUsuarioDto));
+        }
     }
 
     @ResponseStatus(HttpStatus.OK)
@@ -54,7 +70,9 @@ public class UsuarioController
     @PutMapping("/updateUser")
     public int updateUsuarioPassword(@RequestBody String contrasena, Long usuarioId){
         try {
+            if(usuarioRepository.existsById(usuarioId)){
             return usuarioService.setupdateUserPassword(contrasena, usuarioId);
+            }
         } catch (UsuarioGestorExceptions UsuarioGestorExceptions) {
             UsuarioGestorExceptions.printStackTrace();
         }
@@ -66,7 +84,8 @@ public class UsuarioController
     @PutMapping("/updateCorreo")
     public int updateUsuarioCorreo(@RequestBody String correo, Long usuarioId){
         try {
-            return usuarioService.setUpdateUserCorreo(correo, usuarioId);
+            if(usuarioRepository.existsById(usuarioId)){
+            return usuarioService.setUpdateUserCorreo(correo, usuarioId);}
         } catch (UsuarioGestorExceptions UsuarioGestorExceptions) {
             UsuarioGestorExceptions.printStackTrace();
         }
@@ -78,16 +97,29 @@ public class UsuarioController
     @GetMapping("/LoginUser")
     public UsuarioGestorResponse<UsuarioDto> LoginAcces(String nombre, String contrasena)
             throws UsuarioGestorExceptions{
+        if(usuarioRepository.existsUsuarioByNombreAndContrasena(nombre,contrasena))
+        {
         return new UsuarioGestorResponse<>("Succes Login",String.valueOf(HttpStatus.OK),"OK",
-                usuarioService.LoginAcess(nombre,contrasena));
+                usuarioService.LoginAcess(nombre,contrasena));}
+        else
+            {
+                return new UsuarioGestorResponse<>("Fallo al Iniciar Sesión",String.valueOf(HttpStatus.BAD_REQUEST),"Usuario o Contraseña erroneos vuelva a intentar");
+            }
     }
+
     @ResponseStatus(HttpStatus.OK)
     @ApiOperation("Obtiene la lista de notas que posee un Usuario.")
     @GetMapping("/NotasUser")
     public UsuarioGestorResponse<List<NotaDTO>> getNotas(String userId)
             throws UsuarioGestorExceptions{
-        return new UsuarioGestorResponse<>("Succes Login",String.valueOf(HttpStatus.OK),"OK",
+        if(usuarioRepository.existsById(Long.valueOf(userId))){
+        return new UsuarioGestorResponse<>("Success se obtuvo las notas del Usuario",String.valueOf(HttpStatus.OK),"OK",
                 usuarioService.getNotasByUser(userId));
+        }else
+        {
+            return new UsuarioGestorResponse<>("Fallo al obtener notas",String.valueOf(HttpStatus.NOT_FOUND),"Usuario no existe");
+        }
+
     }
 
     @ResponseStatus(HttpStatus.OK)
@@ -96,7 +128,9 @@ public class UsuarioController
     public void deleteUser(Long userId)
     {
         try {
+            if(usuarioRepository.existsById(userId)){
             usuarioService.deleteUser(userId);
+            }
         } catch (UsuarioGestorExceptions UsuarioGestorExceptions) {
             UsuarioGestorExceptions.printStackTrace();
         }
